requirements.txt
requests
pandas
python-dotenv
tensorflow
tensorflow-hub
openai
scikit-learn
nltk
transformers
python-dotenv
